refactor(interactive-section): extract useSectionPersistence hook [Pattern J contract-surface]
Phase: Tier B
Step: B5
Pattern: J (contract-surface ownership move — the largest in this refactor)
Risk: medium-high

Why: The largest Pattern J move in the refactor. The hook now owns
all storage IO + the `interactive-progress-saved` CustomEvent
dispatch + the mount-only RESTORE effect. Four storage namespaces
move ownership in one go:
  - `interactiveStepStorage` (steps per section)
  - `interactiveCompletionStorage` (document-wide percentage)
  - `sectionAcknowledgementStorage` (#842 ack flag)
  - `sectionCollapseStorage` (clear from the reset path only;
    write ownership stays in useSectionAutoCollapse)

API:
  - `persistCompletedSteps(ids)` — writes step set + completion
    percentage, then dispatches `interactive-progress-saved`.
    Preview-mode-gated for writes; the event fires regardless.
  - `clearStepAcknowledgement()` — used by handleStepReset to
    mirror the reducer's "ack cleared with completion" invariant
    in storage.
  - `setAcknowledgement()` — used by handleMarkSectionComplete.
  - `clearAllStorage()` — used by handleResetSection. Sweeps the
    three section-scoped namespaces in one call.
  - The mount-only RESTORE effect runs automatically; calls
    `restoreFromStorage()` then dispatches RESTORE + (on migration)
    `sectionAcknowledgementStorage.set(true)`.

Behaviour pinned by Phase 0:
  - interactive-section.contracts.tripwire.test.tsx all 8 cases
    pass — payload shapes for `interactive-progress-saved`,
    `interactive-progress-cleared`, `section-completed`,
    `interactive-section-completed`, `interactive-step-completed`,
    `pathfinder-step-progress`, plus preview-mode sandbox.
  - interactive-section.state-machine.test.tsx all 9 cases pass —
    end-to-end mount restore + migration + reset + ack flow.
  - data-attributes.contract.test.tsx all cases pass.

The `dispatch` reference (from `useReducer`) is passed into the
hook so the RESTORE effect can dispatch without a closure on the
section component. React guarantees `dispatch` stability across
renders.

Checkpoint:
  - npm run typecheck: pass
  - npx eslint: pass (after pruning two unnecessary deps revealed
    by the rule once the inline code was removed).
  - npx prettier --check: pass
  - npx jest src/components/interactive-tutorial: 19 suites pass,
    294 tests pass + 5 todos + 12 pre-existing skips.
  - interactive-section.tsx line delta: 1528 → 1426 lines (-102).
    Largest single Tier B reduction.
  - All four storage imports drop out of interactive-section.tsx.
  - `restoreFromStorage` import drops out too (it moves into the
    hook).
  - `getTotalDocumentSteps` import drops out — only the hook reads
    it now.

Notes:
  - The `try/catch` with no-op catch around `window.dispatchEvent`
    is NOT preserved in the hook because the pre-extraction code
    didn't have it for the `interactive-progress-saved` dispatch
    (only for `pathfinder-step-progress`, which lives in B3).
    Verified by re-reading the original; no semantic change.
  - The dispatch identity from `useReducer` is stable across
    renders, so passing it into the hook doesn't churn the hook's
    effect deps.
  - `clearStepAcknowledgement` is named for its semantic role
    ("acknowledgement clears alongside completion"), not for the
    storage namespace, so a future namespace rename wouldn't
    cascade through call sites.
